+++ src/app/BUILD.gn
@@ -124,8 +124,8 @@ config("config-controller-dynamic-server") {
 
 source_set("constants") {
   sources = [
-    "InteractionModelRevision.h",
     "InteractionModelTimeout.h",
+    "SpecificationDefinedRevisions.h",
   ]
   public_deps = [
     "${chip_root}/src/lib/core",
