+++ src/app/MessageDef/MessageBuilder.h
@@ -18,7 +18,7 @@
 #pragma once
 
 #include "StructBuilder.h"
-#include <app/InteractionModelRevision.h>
+#include <app/SpecificationDefinedRevisions.h>
 #include <app/util/basic-types.h>
 
 namespace chip {
