+++ src/app/MessageDef/MessageDefHelper.cpp
@@ -24,7 +24,7 @@
 #include "MessageDefHelper.h"
 #include <algorithm>
 #include <app/AppConfig.h>
-#include <app/InteractionModelRevision.h>
+#include <app/SpecificationDefinedRevisions.h>
 #include <app/util/basic-types.h>
 #include <inttypes.h>
 #include <lib/support/logging/CHIPLogging.h>
